Release and Deploy mdTools v1.1.1 (#69)
* GitHub Action to Deploy to GitHub Pages (and updates) (#58)

* Upgrade mdCodes and mdJson-schemas

* Increment Patch Version

* Update version

* add workflow

* Set version to 1.1.2

---------

Co-authored-by: Jonathan Oliveros <joliveros@xentity.com>

* Fix GitHub Actions Trigger (#60)

Removing the requirement that the action only run when a PR is merged. This overcomplicates the process for using the token to write back to the gh-pages repository and is not needed when there are branch protections requiring that all merges must come from a PR.

* Chore/update versions (#62)

* Fix GitHub Actions Trigger

Removing the requirement that the action only run when a PR is merged. This overcomplicates the process for using the token to write back to the gh-pages repository and is not needed when there are branch protections requiring that all merges must come from a PR.

* Bump mdcodes and mdjson-schemas to version 2.10.0 and 2.9.5 respectively

* Refactor GitHub Actions workflow to include build and deploy jobs, update action versions, and automate tagging and release creation.

* Bump version to 1.1.3 in package.json and package-lock.json

* Downgrade version to 1.1.1 in package.json

---------

Co-authored-by: Jonathan Oliveros <joliveros@xentity.com>